requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
openai==1.12.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
gunicorn==21.2.0 